Issue #2994506 Fix tests to adjust to new class hierarchy
